Split IPTT ReportView tests into their own file
